cancel registration selected dancers

+++ run.py
@@ -1,8 +1,9 @@
 from ntds_webportal import create_app, db
 import sqlalchemy as alchemy
 # noinspection PyPackageRequirements
-from instance.populate import populate_test_db, create_users
-from ntds_webportal.models import User, Notification
+from instance.populate import populate_test_db, create_users, create_tournament
+from ntds_webportal.models import User, Notification, EXCLUDED_FROM_CLEARING, TeamFinances
+
 
 app = create_app()
 
@@ -25,6 +26,7 @@ class DevShell:
         with app.app_context():
             print('Creating Users.')
             create_users()
+            create_tournament()
 
     @staticmethod
     def populate_test():
@@ -37,15 +39,25 @@ class DevShell:
         with app.app_context():
             meta = db.metadata
             for table in reversed(meta.sorted_tables):
-                print('Cleared table {}.'.format(table))
-                db.session.execute(table.delete())
+                if table.name not in EXCLUDED_FROM_CLEARING:
+                    print('Cleared table {}.'.format(table))
+                    db.session.execute(table.delete())
+            tf = TeamFinances.query.all()
+            for f in tf:
+                f.paid = 0
+            create_tournament()
             db.session.commit()
 
     @staticmethod
-    def reset(self):
-        self.clear()
-        self.create()
-        self.populate_test()
+    def reset_test_data(self):
+        with app.app_context():
+            meta = db.metadata
+            for table in reversed(meta.sorted_tables):
+                print('Cleared table {}.'.format(table))
+                db.session.execute(table.delete())
+            db.session.commit()
+            self.create()
+            self.populate_test()
 
 
 if __name__ == '__main__':
